CHORE: cleaned tdf blob reader
